Use d3.geo.path and SVG path element to render features
Convert features to GeoJSON and use those objects for the d3 data
join instead of our proprietary format.  Also use an SVG path element
instead of circles/lines for displaying facilities and transfers.
In order to do this, we use a d3.geo.path instance when setting the
feature elements ``d`` attribute.  This greatly simplifies the
code for setting the position attributes of the SVG elements and
also eliminates the need to have our own code to translate the lat/lon
coordinates for our projection/scaling.